+++ master/docs/release-notes.rst
@@ -108,6 +108,9 @@ Changes for Developers
   corresponding (undocumented)
   :py:class:`buildbot.db.buildsets.BuildsetsConnector` methods.
 
+* Errors during configuration (in particular in :py:class:`BuildStep` constructors),
+  should be reported by calling :py:function:`buildbot.config.error`.
+
 Features
 ~~~~~~~~